refactor(web): move getMcpOAuthCallbackUrl into mcp utils.server
