+++ packages/web/src/app/(app)/settings/workspaceAskAgent/page.tsx
@@ -1,7 +1,7 @@
 import { hasEntitlement } from "@/lib/entitlements";
 import { authenticatedPage } from "@/middleware/authenticatedPage";
 import { OrgRole } from "@sourcebot/db";
-import { getMcpOAuthCallbackUrl } from "@/ee/features/chat/mcp/mcpOAuthCallbackUrl";
+import { getMcpOAuthCallbackUrl } from "@/ee/features/chat/mcp/utils.server";
 import { WorkspaceAskAgentPage } from "./workspaceAskAgentPage";
 import { WorkspaceAskAgentEntitlementMessage } from "./workspaceAskAgentEntitlementMessage";
